+++ bytebankHerdado/src/bytebankHerdado/Administrador.java
@@ -1,0 +1,10 @@
+package bytebankHerdado;
+
+public class Administrador extends FuncionarioAutenticavel {
+	
+	@Override
+	public double getBonificacao() {
+		return 50;
+	}
+
+}
