+++ bytebankHerdado/src/bytebankHerdado/FuncionarioAutenticavel.java
@@ -1,0 +1,19 @@
+package bytebankHerdado;
+
+public abstract class FuncionarioAutenticavel extends Funcionario {
+
+	private int senha;
+	
+	public void setSenha(int senha) {
+		this.senha = senha;
+	}
+	
+	public boolean autentica(int senha) {
+		if (this.senha == senha) {
+			return true;
+		} else {
+			return false;
+		}
+	}
+
+}
