+++ bytebankHerdado/src/bytebankHerdado/Gerente.java
@@ -1,21 +1,7 @@
 package bytebankHerdado;
 
-//Gerente é um Funcionário ou Gerente herda da classe Funcionário
-public class Gerente extends Funcionario {
-
-	private int senha;
-	
-	public void setSenha(int senha) {
-		this.senha = senha;
-	}
-	
-	public boolean autentica(int senha) {
-		if (this.senha == senha) {
-			return true;
-		} else {
-			return false;
-		}
-	}
+//Gerente é um FuncionarioAutenticavel ou Gerente herda da classe FuncionarioAutenticavel
+public class Gerente extends FuncionarioAutenticavel {
 	
 	public double getBonificacao() {
 		System.out.println("Chamando o método de bonificação do GERENTE");
