+++ bytebankHerdado/src/bytebankHerdado/SistemaInterno.java
@@ -1,0 +1,17 @@
+package bytebankHerdado;
+
+public class SistemaInterno {
+	
+	private int senha = 2222;
+	
+	public void autentica(FuncionarioAutenticavel fa) {
+		boolean autenticou = fa.autentica(this.senha);
+		
+		if (autenticou) {
+			System.out.println("Pode entrar no sistema!");
+		} else {
+			System.out.println("Não pode entrar no sistema!");
+		}
+	}
+
+}
