+++ bytebankHerdado/src/bytebankHerdado/TesteSistema.java
@@ -1,0 +1,18 @@
+package bytebankHerdado;
+
+public class TesteSistema {
+
+	public static void main(String[] args) {
+
+		Gerente g = new Gerente();
+		g.setSenha(2222);
+		
+		Administrador adm = new Administrador();
+		adm.setSenha(3333);
+		
+		SistemaInterno si =  new SistemaInterno();
+		si.autentica(g);
+		si.autentica(adm);
+	}
+
+}
